reuse retrieveBreadcrumbs, move to util
