+++ src/containers/TopicArticlePage/components/TopicArticleTaxonomy.jsx
@@ -30,6 +30,7 @@ import {
   pathToUrnArray,
 } from '../../../util/taxonomyHelpers';
 import handleError from '../../../util/handleError';
+import retriveBreadCrumbs from '../../../util/retriveBreadCrumbs';
 import SaveButton from '../../../components/SaveButton';
 import { FormikActionButton } from '../../FormikForm';
 import TopicArticleConnections from './TopicArticleConnections';
@@ -50,22 +51,13 @@ class TopicArticleTaxonomy extends Component {
         allTopics: [],
       },
     };
-    this.retriveBreadCrumbs = this.retriveBreadCrumbs.bind(this);
-    this.stageTaxonomyChanges = this.stageTaxonomyChanges.bind(this);
-    this.removeConnection = this.removeConnection.bind(this);
-    this.getSubjectTopics = this.getSubjectTopics.bind(this);
-    this.updateSubject = this.updateSubject.bind(this);
-    this.handleSubmit = this.handleSubmit.bind(this);
-    this.fetchTaxonomy = this.fetchTaxonomy.bind(this);
-    this.createAndPlaceTopic = this.createAndPlaceTopic.bind(this);
-    this.onCancel = this.onCancel.bind(this);
   }
 
   componentDidMount() {
     this.fetchTaxonomy();
   }
 
-  async getSubjectTopics(subjectId) {
+  getSubjectTopics = async subjectId => {
     if (
       this.state.structure.some(
         subject => subject.id === subjectId && subject.topics,
@@ -81,9 +73,9 @@ class TopicArticleTaxonomy extends Component {
     } catch (e) {
       handleError(e);
     }
-  }
+  };
 
-  async fetchTaxonomy() {
+  fetchTaxonomy = async () => {
     const {
       article: { language, id },
     } = this.props;
@@ -117,9 +109,9 @@ class TopicArticleTaxonomy extends Component {
       handleError(e);
       this.setState({ status: 'error' });
     }
-  }
+  };
 
-  stageTaxonomyChanges({ addTopicId, removeTopicId, path }) {
+  stageTaxonomyChanges = ({ addTopicId, removeTopicId, path }) => {
     const {
       article: { title },
     } = this.props;
@@ -154,9 +146,9 @@ class TopicArticleTaxonomy extends Component {
         ],
       }));
     }
-  }
+  };
 
-  async createAndPlaceTopic(topic, articleId) {
+  createAndPlaceTopic = async (topic, articleId) => {
     const newTopicPath = await addTopic({
       name: topic.name,
       contentUri: `urn:article:${articleId}`,
@@ -181,9 +173,9 @@ class TopicArticleTaxonomy extends Component {
       id: newTopicId,
       path: topic.path.replace('staged', newTopicId.replace('urn:', '')),
     };
-  }
+  };
 
-  async handleSubmit(evt) {
+  handleSubmit = async evt => {
     evt.preventDefault();
     const { stagedTopicChanges, topics } = this.state;
     const {
@@ -246,9 +238,9 @@ class TopicArticleTaxonomy extends Component {
       handleError(err);
       this.setState({ saveStatus: 'error' });
     }
-  }
+  };
 
-  updateSubject(subjectid, newSubject) {
+  updateSubject = (subjectid, newSubject) => {
     this.setState(prevState => ({
       structure: prevState.structure.map(subject => {
         if (subject.id === subjectid) {
@@ -257,50 +249,13 @@ class TopicArticleTaxonomy extends Component {
         return subject;
       }),
     }));
-  }
+  };
 
-  retriveBreadCrumbs(topicPath) {
-    const {
-      structure,
-      taxonomyChoices: { allTopics },
-    } = this.state;
-    const {
-      article: { title },
-    } = this.props;
-    try {
-      const [subjectPath, ...topicPaths] = pathToUrnArray(topicPath);
-
-      const subject = structure.find(
-        structureSubject => structureSubject.id === subjectPath,
-      );
-      const returnPaths = [];
-      returnPaths.push({
-        name: subject.name,
-        id: subject.id,
-      });
-      topicPaths.forEach(pathId => {
-        const topicPath = allTopics.find(subtopic => subtopic.id === pathId);
-        if (topicPath) {
-          returnPaths.push({
-            name: topicPath.name,
-            id: topicPath.id,
-          });
-        } else {
-          returnPaths.push({ name: title, id: pathId });
-        }
-      });
-      return returnPaths;
-    } catch (err) {
-      handleError(err);
-      return false;
-    }
-  }
-
-  removeConnection(id) {
+  removeConnection = id => {
     this.stageTaxonomyChanges({ removeTopicId: id });
-  }
+  };
 
-  onCancel() {
+  onCancel = () => {
     const { isDirty } = this.state;
     const { closePanel } = this.props;
     if (!isDirty) {
@@ -309,7 +264,7 @@ class TopicArticleTaxonomy extends Component {
       // TODO open warning
       closePanel();
     }
-  }
+  };
 
   render() {
     const {
@@ -320,7 +275,10 @@ class TopicArticleTaxonomy extends Component {
       saveStatus,
       isDirty,
     } = this.state;
-    const { t } = this.props;
+    const {
+      t,
+      article: { title },
+    } = this.props;
 
     if (status === 'loading') {
       return <Spinner />;
@@ -347,7 +305,9 @@ class TopicArticleTaxonomy extends Component {
           structure={structure}
           taxonomyTopics={allTopics}
           activeTopics={stagedTopicChanges}
-          retriveBreadCrumbs={this.retriveBreadCrumbs}
+          retriveBreadCrumbs={topicPath =>
+            retriveBreadCrumbs({ topicPath, allTopics, structure, title })
+          }
           removeConnection={this.removeConnection}
           getSubjectTopics={this.getSubjectTopics}
           stageTaxonomyChanges={this.stageTaxonomyChanges}
